backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
crewai==1.9.3
litellm>=1.75.0
langgraph>=0.0.35
langchain-groq>=0.1.3
langchain-core>=0.1.30
python-dotenv>=1.0.1
pydantic>=2.5.0
typing-extensions>=4.11.0
